refactors mounting and adds tests
