+++ pkg/machine/ignition.go
@@ -10,8 +10,6 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 
 	"github.com/containers/common/pkg/config"
 	"github.com/sirupsen/logrus"
@@ -218,10 +216,10 @@ WantedBy=sysinit.target
 			Contents: &envset,
 		},
 	}
-	mountUnits := buildMountUnits(ign)
-	logrus.Trace("Appending mount systemd units")
-	units = append(units, mountUnits...)
-	logrus.Tracef("%+v", units)
+	if ign.Mounts != nil {
+		mountUnits := buildMountUnits(ign.Mounts)
+		units = append(units, mountUnits...)
+	}
 	ignSystemd := Systemd{Units: units}
 
 	ignConfig := Config{
@@ -630,91 +628,4 @@ func getLinks(usrName string) []Link {
 
 func encodeDataURLPtr(contents string) *string {
 	return strToPtr(fmt.Sprintf("data:,%s", url.PathEscape(contents)))
-}
-
-func buildMountUnits(ign DynamicIgnition) []Unit {
-	mountUnits := []Unit{}
-	systemdMountTemplate := `[Unit]
-Description=UserVolume Mount
-Restart=on-failure
-
-[Mount]
-What=%s
-Where=%s
-Type=%s
-Options=defaults
-DirectoryMode=0755
-
-[Install]
-WantedBy=default.target`
-
-	volumeMountPointTemplate := `[Unit]
-Description=Ensures %s Mountpoint
-Before=%s
-
-[Service]
-ExecStart=sh -c 'chattr -i / ; mkdir -p %s ; chattr +i / ;'
-
-[Install]
-WantedBy=default.target`
-
-	if ign.Mounts == nil {
-		return mountUnits
-	}
-
-	for _, mount := range ign.Mounts {
-		systemdName := fsPathToMountName(mount.Target)
-		mountUnitName := fmt.Sprintf("%s.mount", systemdName)
-		serviceUnitName := fmt.Sprintf("%s.service", systemdName)
-
-		if !strings.HasPrefix(mount.Target, "/home") && !strings.HasPrefix(mount.Target, "/mnt") {
-			mkdirUnit := Unit{
-				Enabled: boolToPtr(true),
-				Name:    serviceUnitName,
-				Contents: strToPtr(fmt.Sprintf(
-					volumeMountPointTemplate,
-					mount.Tag,
-					mountUnitName,
-					mount.Target,
-				)),
-			}
-
-			mountUnits = append(mountUnits, mkdirUnit)
-		}
-
-		unit := Unit{
-			Enabled: boolToPtr(true),
-			Name:    mountUnitName,
-			Contents: strToPtr(fmt.Sprintf(
-				systemdMountTemplate,
-				mount.Tag,
-				mount.Target,
-				mount.Type,
-			)),
-		}
-
-		mountUnits = append(mountUnits, unit)
-	}
-	return mountUnits
-}
-
-func fsPathToMountName(path string) string {
-	// according to the systemd docs we need
-	// to replace the slashes with dashes
-	// and drop proceeding slashes so
-	// /var/lib/podman would become
-	// var-lib-podman
-	if path[0:1] == "/" {
-		path = path[1:]
-	}
-	// also lets drop any trailing slashes
-	if path[len(path)-1:] == "/" {
-		path = path[0 : len(path)-1]
-	}
-
-	// replace with dashes
-	re := regexp.MustCompile(`/+`)
-	path = re.ReplaceAllString(path, "-")
-
-	return path
 }
